requirements.txt
backoff==2.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
Deprecated==1.2.13
googleapis-common-protos==1.56.2
grpcio==1.50.0
idna==3.4
opentelemetry-api==1.13.0
opentelemetry-exporter-otlp==1.13.0
opentelemetry-exporter-otlp-proto-grpc==1.13.0
opentelemetry-exporter-otlp-proto-http==1.13.0
opentelemetry-proto==1.13.0
opentelemetry-sdk==1.13.0
opentelemetry-semantic-conventions==0.34b0
protobuf==3.20.3
requests==2.28.1
six==1.16.0
thrift==0.16.0
typing_extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
Deprecated==1.2.13
opentelemetry-api==1.12.0
opentelemetry-sdk==1.12.0
opentelemetry-semantic-conventions==0.33b0
typing_extensions==4.3.0
wrapt==1.14.1
